requirements.txt
fastapi==0.115.0
gseapy==1.1.3
matplotlib==3.8.2
networkx==3.2.1
numpy>=1.21,<2
pandas==2.2.3
pydantic==2.9.2
PyYAML==6.0.2
scipy==1.14.1
starlette>=0.37.2,<0.39.0
statsmodels==0.14.2
uvicorn==0.25.0
python-multipart==0.0.6
openpyxl==3.1.2
jinja2==3.1.2
anyio>=4.9.0
idna>=3.10
sniffio>=1.3.1
python-dateutil>=2.9.0
pytz>=2025.2
python-tzdata>=2025.2
